Exploded collection models to separate files. Added recommendable field
