+++ workflows/api/serializers/workflows/collection.py
@@ -101,6 +101,7 @@ class WorkflowCollectionSummarySerializer(WorkflowCollectionBaseSerializer):
             'home_image',
             'library_image',
             'assignment_only',
+            'recommendable',
             'name',
             'ordered',
             'authors',
@@ -136,6 +137,7 @@ class WorkflowCollectionDetailedSerializer(WorkflowCollectionBaseSerializer):
             'home_image',
             'library_image',
             'assignment_only',
+            'recommendable',
             'name',
             'ordered',
             'workflowcollectionmember_set',
@@ -172,6 +174,7 @@ class WorkflowCollectionWithStepsSerializer(WorkflowCollectionBaseSerializer):
             'home_image',
             'library_image',
             'assignment_only',
+            'recommendable',
             'name',
             'ordered',
             'workflowcollectionmember_set',
