+++ workflows/api/tests/views/workflows/test_collection.py
@@ -68,6 +68,7 @@ class TestWorkflowCollectionsView(TestCase):
                     "home_image",
                     "library_image",
                     "assignment_only",
+                    "recommendable",
                     "name",
                     "ordered",
                     "authors",
@@ -151,6 +152,7 @@ class TestWorkflowCollectionView(TestCase):
                 "home_image",
                 "library_image",
                 "assignment_only",
+                "recommendable",
                 "name",
                 "ordered",
                 "workflowcollectionmember_set",
@@ -215,6 +217,7 @@ class TestWorkflowCollectionView(TestCase):
                 "home_image",
                 "library_image",
                 "assignment_only",
+                "recommendable",
                 "name",
                 "ordered",
                 "workflowcollectionmember_set",
